+++ view/frontend/web/js/view/payment/nuvei.js
@@ -25,7 +25,7 @@ define(
                 : 'https://cdn.safecharge.com/safecharge_resources/v1/checkout/checkout.js';
                 
 		document.head.appendChild(nuveiCheckoutSdkScr);
-		// Load Nuvei Chekout SDK and add it ot a local variable
+		// /Load Nuvei Chekout SDK and add it ot a local variable
 		
         rendererList.push({
 			type: 'nuvei',